+++ packages/orq-rc/src/models/operations/getallprompts.ts
@@ -9,7 +9,25 @@ import { ClosedEnum } from "../../types/enums.js";
 import { Result as SafeParseResult } from "../../types/fp.js";
 import { SDKValidationError } from "../errors/sdkvalidationerror.js";
 
+/**
+ * List sorting preference.
+ */
+export const GetAllPromptsQueryParamSort = {
+  Asc: "asc",
+  Desc: "desc",
+} as const;
+/**
+ * List sorting preference.
+ */
+export type GetAllPromptsQueryParamSort = ClosedEnum<
+  typeof GetAllPromptsQueryParamSort
+>;
+
 export type GetAllPromptsRequest = {
+  /**
+   * List sorting preference.
+   */
+  sort?: GetAllPromptsQueryParamSort | undefined;
   /**
    * A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
    */
@@ -487,11 +505,33 @@ export type GetAllPromptsResponseBody = {
 };
 
 /** @internal */
+export const GetAllPromptsQueryParamSort$inboundSchema: z.ZodNativeEnum<
+  typeof GetAllPromptsQueryParamSort
+> = z.nativeEnum(GetAllPromptsQueryParamSort);
+
+/** @internal */
+export const GetAllPromptsQueryParamSort$outboundSchema: z.ZodNativeEnum<
+  typeof GetAllPromptsQueryParamSort
+> = GetAllPromptsQueryParamSort$inboundSchema;
+
+/**
+ * @internal
+ * @deprecated This namespace will be removed in future versions. Use schemas and types that are exported directly from this module.
+ */
+export namespace GetAllPromptsQueryParamSort$ {
+  /** @deprecated use `GetAllPromptsQueryParamSort$inboundSchema` instead. */
+  export const inboundSchema = GetAllPromptsQueryParamSort$inboundSchema;
+  /** @deprecated use `GetAllPromptsQueryParamSort$outboundSchema` instead. */
+  export const outboundSchema = GetAllPromptsQueryParamSort$outboundSchema;
+}
+
+/** @internal */
 export const GetAllPromptsRequest$inboundSchema: z.ZodType<
   GetAllPromptsRequest,
   z.ZodTypeDef,
   unknown
 > = z.object({
+  sort: GetAllPromptsQueryParamSort$inboundSchema.default("asc"),
   limit: z.number().default(10),
   starting_after: z.string().optional(),
   ending_before: z.string().optional(),
@@ -504,6 +544,7 @@ export const GetAllPromptsRequest$inboundSchema: z.ZodType<
 
 /** @internal */
 export type GetAllPromptsRequest$Outbound = {
+  sort: string;
   limit: number;
   starting_after?: string | undefined;
   ending_before?: string | undefined;
@@ -515,6 +556,7 @@ export const GetAllPromptsRequest$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GetAllPromptsRequest
 > = z.object({
+  sort: GetAllPromptsQueryParamSort$outboundSchema.default("asc"),
   limit: z.number().default(10),
   startingAfter: z.string().optional(),
   endingBefore: z.string().optional(),